Steps to create a package RPM

+++ src/cpf.c
@@ -3,7 +3,7 @@
 char* format_cpf(char* number) {
 	int i, pos;
 	char *result = malloc(15); 
-	for (i=0, pos=0; i<strlen(number); i++) {
+	for (i=0, pos=0; i<(int)strlen(number); i++) {
 		if (i == 3 || i == 6) {
 			result[pos++] = '.';
 		} else if (i == 9) {
